requirements.txt
Flask
flask-bcrypt
Flask-JWT-Extended
Flask-Migrate
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
